+++ api/core/doc.go
@@ -1,0 +1,2 @@
+// package core is empty
+package core