+++ api/core/v3/validation.go
@@ -1,0 +1,23 @@
+package v3
+
+import (
+	"errors"
+
+	corev2 "github.com/sensu/sensu-go/api/core/v2"
+)
+
+func ValidateMetadata(meta *corev2.ObjectMeta) error {
+	if meta == nil {
+		return errors.New("nil metadata")
+	}
+	if err := corev2.ValidateName(meta.Name); err != nil {
+		return err
+	}
+	if meta.Labels == nil {
+		return errors.New("nil labels")
+	}
+	if meta.Annotations == nil {
+		return errors.New("nil annotations")
+	}
+	return nil
+}