+++ CHANGELOG.md
@@ -10,8 +10,20 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### Added
 - Added `SENSU_LICENSE_FILE` environment variable to handler executions.
+- Added the `sensuctl user hash-password` command to generate password hashes
+- Added the ability to reset passwords via the backend API & `sensuctl user
+reset-password`
+- The user resource now has a `password_hash` field to specify the password
+hash, in place of the cleartext password
+- Added the beginnings of the core/v3 API.
+- Added automatically generated tests for the core/v2 API.
+
+### Changed
+- `sensu-agent`'s default log level is now `info` instead of `warn`.
 
 ### Fixed
+- The password verification logic when running `sensuctl user change-password`
+has been moved from sensuctl to the backend API.
 - Errors while publishing proxy check requests do not block scheduling for other
 entities.
 - Listing namespaces in sensuctl with `--chunk-size` now works properly.
@@ -34,6 +46,12 @@ they were intended to be.
 ## [5.20.1] - 2020-05-15
 *No changelog for this release.*
 
+### Added
+- Added the ability to omit types from sensuctl dump when using the "all" flag.
+
+### Fixed
+- Fixed a bug in sensuctl dump where duplicate resource names could be specified.
+
 ## [5.20.0] - 2020-05-12
 
 ### Added
@@ -47,6 +65,8 @@ sensu-backend or sensu-agent process.
 - Added token substitution for assets.
 - Added `Edition` field to version information.
 - Added `GoVersion` field to version information.
+- Assets paths are now accessible to consumers via an environment variable.
+- Added a helper function to token substitution to retrieve an asset path.
 - Windows agent now has log rotation capabilities.
 - [Web] Added check hook output to event details page.